refactor: move InfoBox out of the search-result component as should not be used in advanced search
